Arreglo de Token

+++ GlassCoreWebAPI/GlassCoreWebAPI/Models/DTOs/UsuarioDTOs/LoginUsuarioDTO.cs
@@ -2,7 +2,7 @@
 {
     public class LoginUsuarioDTO
     {
-        public long UserName { get; set; }
+        public int UserName { get; set; }
         public string Password { get; set; } = null!;
